Refactor cache set up out of system-initialise module
Ticket: #107
