+++ src/modules/orders/index.ts
@@ -1,3 +1,4 @@
 import { logger } from '../../system/logging';
 import './use-cases';
+
 logger.info('Parse Cloud: Orders module loaded.');
