+++ src/modules/orders/mappers/index.ts
@@ -1,3 +1,5 @@
 export { CustomerPersistenceMapper } from './customer-persistence.mapper';
 export { OrderDTOMapper as SubmissionDTOMapper } from './order-dto.mapper';
 export { SampleEntryDTOMapper } from './sample-entry-dto.mapper';
+
+export { NRLPersistenceMapper } from './nrl-persistence.mapper';
