+++ src/modules/orders/use-cases/index.ts
@@ -1,4 +1,12 @@
 import './create-submission-file';
 import './parse-sample-data';
+import './set-avv-catalogue-cache';
+import './set-nrl-cache';
+import './set-plz-cache';
+import './set-search-alias-cache';
 import './submit-order';
 import './validate-order';
+
+export { setAVVCatalogueCache } from './set-avv-catalogue-cache';
+
+export { setSearchAliasCache } from './set-search-alias-cache';
