+++ src/modules/system-initialise/index.ts
@@ -3,45 +3,14 @@
  * - set the system up to function properly and check system is working as plannned
  */
 import { logger } from '../../system/logging';
-import { ObjectKeys } from '../shared/infrastructure/parse-types';
 import {
-    afterDeleteAVVCatalogueHook,
-    afterDeleteNRLHook,
-    afterDeletePLZHook,
-    afterDeleteSearchAliasHook,
-    afterSaveAVVCatalogueHook,
-    afterSaveNRLHook,
-    afterSavePLZHook,
-    afterSaveSearchAliasHook,
     checkCollectionsForContent,
-    checkSystemConfiguration,
-    setNRLCache,
-    setPLZCache
+    checkSystemConfiguration
 } from './use-cases';
 
 logger.info('Parse Cloud: Checking System Configuration.');
 checkSystemConfiguration.execute();
 logger.info('Parse Cloud: Checking Collections for content.');
 checkCollectionsForContent.execute();
-
-logger.info('Parse Cloud: Creating NRL cache.');
-setNRLCache.execute();
-Parse.Cloud.afterSave(ObjectKeys.NRL, afterSaveNRLHook);
-Parse.Cloud.afterDelete(ObjectKeys.NRL, afterDeleteNRLHook);
-
-logger.info('Parse Cloud: Creating PLZ cache.');
-setPLZCache.execute();
-Parse.Cloud.afterSave(ObjectKeys.AllowedPLZ, afterSavePLZHook);
-Parse.Cloud.afterDelete(ObjectKeys.AllowedPLZ, afterDeletePLZHook);
-
-logger.info('Parse Cloud: Setting AVVCatalogue hooks.');
-// AVVCatalogue Cache is created at src/modules/orders/legacy/index.ts
-Parse.Cloud.afterSave(ObjectKeys.AVVCatalogue, afterSaveAVVCatalogueHook);
-Parse.Cloud.afterDelete(ObjectKeys.AVVCatalogue, afterDeleteAVVCatalogueHook);
-
-logger.info('Parse Cloud: Creating Search-Alias hooks.');
-// Search-Alias Cache is created at src/modules/orders/legacy/index.ts
-Parse.Cloud.afterSave(ObjectKeys.SearchAlias, afterSaveSearchAliasHook);
-Parse.Cloud.afterDelete(ObjectKeys.SearchAlias, afterDeleteSearchAliasHook);
 
 logger.info('Parse Cloud: System initialise module loaded.');
