--- src/modules/system-initialise/mappers/index.ts
@@ -1,1 +1,0 @@
-export { NRLPersistenceMapper } from './nrl-persistence.mapper';
