+++ src/modules/system-initialise/use-cases/index.ts
@@ -1,24 +1,2 @@
 export { checkSystemConfiguration } from './checkSystemConfiguration';
-export {
-    afterDeleteAVVCatalogueHook,
-    afterSaveAVVCatalogueHook,
-    setAVVCatalogueCache
-} from './setAVVCatalogueCache';
-export {
-    afterDeleteNRLHook,
-    afterSaveNRLHook,
-    setNRLCache
-} from './setNRLCache';
-export {
-    afterDeletePLZHook,
-    afterSavePLZHook,
-    setPLZCache
-} from './setPLZCache';
-
 export { checkCollectionsForContent } from './checkCollectionsForContent';
-
-export {
-    afterDeleteSearchAliasHook,
-    afterSaveSearchAliasHook,
-    setSearchAliasCache
-} from './setSearchAliasCache';
