--- src/modules/system-initialise/use-cases/setAVVCatalogueCache/index.ts
@@ -1,8 +1,0 @@
-export {
-    afterDeleteAVVCatalogueHook,
-    afterSaveAVVCatalogueHook
-} from './set-avvcatalogue-cache.hook';
-export {
-    SetAVVCatalogueCacheUseCase,
-    setAVVCatalogueCache
-} from './set-avvcatalogue-cache.use-case';
