--- src/modules/system-initialise/use-cases/setNRLCache/index.ts
@@ -1,2 +1,0 @@
-export { afterDeleteNRLHook, afterSaveNRLHook } from './set-nrl-cache.hook';
-export { SetNRLCacheUseCase, setNRLCache } from './set-nrl-cache.use-case';
