--- src/modules/system-initialise/use-cases/setPLZCache/index.ts
@@ -1,2 +1,0 @@
-export { afterDeletePLZHook, afterSavePLZHook } from './set-plz-cache.hook';
-export { SetPLZCacheUseCase, setPLZCache } from './set-plz-cache.use-case';
