--- src/modules/system-initialise/use-cases/setSearchAliasCache/index.ts
@@ -1,8 +1,0 @@
-export {
-    afterDeleteSearchAliasHook,
-    afterSaveSearchAliasHook
-} from './set-search-alias-cache.hook';
-export {
-    SetSearchAliasCacheUseCase,
-    setSearchAliasCache
-} from './set-search-alias-cache.use-case';
